GH-16278: add has_multi_thread flag to as_data_frame and adjust tests.
