+++ h2o-py/tests/pyunit_utils/__init__.py
@@ -3,3 +3,4 @@ from .utils_model_metrics import *
 from .utils_model_custom_distribution import *
 from .utils_for_glm_tests import *
 from .sklearn_multinomial_auc_method import roc_auc_score
+from .utils_parser_tests import *
